refactor: introduce store.js façade over appState (roadmap step 4)
Add src/store.js exporting appState, getState, setState, and subscribe.
state.js becomes a backwards-compatibility re-export. setState mutates
the existing appState via Object.assign so direct `appState.x = y`
writes from parser/renderer/controls/midi/main keep hitting the same
object the façade reads from. No caller is migrated yet — Step 5 will
move parser.js writes through setState.

Deduped currentSoftkeys, which appeared twice in the source literal
(no runtime shape change — JS keeps the second).
